AI Agent - Azure Codes/Demo1/requirements.txt
azure-ai-projects>=2.0.0a20250915020
azure-identity
azure-ai-agents
python-dotenv